src: Refactors. -Rename modules to PEP standard. -Rename Utils module to tools.
