+++ src/actuator/actuator.py
@@ -1,4 +1,4 @@
-from Utils.status import Status
+from tools.status import Status
 from pubsub import pub
 # Actuator Interface
